Fix dmd new version issues

+++ source/netload/protocols/dns/dns.d
@@ -697,7 +697,7 @@ class DNSQR : Protocol {
     }
 
     override ubyte[] toBytes() const {
-      ulong inc = (_qname.length > 1 ? _qname.length + 1 : 0);
+      uint inc = (_qname.length > 1 ? _qname.length + 1 : 0);
       ubyte[] packet = new ubyte[5 + inc];
 
       writeLabels(0, _qname, packet);
@@ -862,7 +862,7 @@ class DNSRR : Protocol {
     }
 
     override ubyte[] toBytes() const {
-      ulong inc = (_rname.length > 1 ? _rname.length + 1 : 0);
+      uint inc = (_rname.length > 1 ? _rname.length + 1 : 0);
       ubyte[] packet = new ubyte[11 + inc];
       writeLabels(0, _rname, packet);
       packet.write!ushort(_rtype, (1 + inc));
@@ -1059,7 +1059,7 @@ class DNSSOAResource  : Protocol {
     }
 
     override ubyte[] toBytes() const {
-      ulong inc = (_primary.length > 1 ? _primary.length + 1 : 0) + (_admin.length > 1 ? _admin.length + 1 : 0);
+      uint inc = (_primary.length > 1 ? _primary.length + 1 : 0) + (_admin.length > 1 ? _admin.length + 1 : 0);
       ubyte[] packet = new ubyte[22 + inc];
       uint pos = 0;
 
@@ -1243,7 +1243,7 @@ class DNSMXResource : Protocol {
     }
 
     override ubyte[] toBytes() const {
-      ulong inc = (_mxname.length > 1 ? _mxname.length + 1 : 0);
+      uint inc = (_mxname.length > 1 ? _mxname.length + 1 : 0);
       ubyte[] packet = new ubyte[3 + inc];
       packet.write!ushort(_pref, 0);
       writeLabels(2, _mxname, packet);
@@ -1487,7 +1487,7 @@ class DNSPTRResource : Protocol {
     }
 
     override ubyte[] toBytes() const {
-      ulong inc = (_ptrname.length > 1 ? _ptrname.length + 1 : 0);
+      uint inc = (_ptrname.length > 1 ? _ptrname.length + 1 : 0);
       ubyte[] packet = new ubyte[1 + inc];
       writeLabels(0, _ptrname, packet);
       if (_data !is null)
